requirements.txt
#backend
fastapi==0.115.0
# langchain
# langchain-google-community
# google-cloud-bigquery
# langchain-google-genai
# langchain_google_vertexai
# vertexai
# pandas
# langchain-google-community[featurestore]
# beautifulsoup4
# docx2txt==0.8
# pypdf==5.0.0




#pipeline
langchain-google-community[gcs]
langchain-google-vertexai==1.0.8
google-cloud-aiplatform==1.60.0
google-cloud-discoveryengine==0.12.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.2
docx2txt==0.8
pypdf==5.0.0
pandas==2.2.3
python-dotenv==1.0.1
pyarrow
apache-beam[gcp]

#GCSloader
unstructured[docx]
python-docx
python-magic-bin